3989 - move users/index content to users/users and export
